+++ core/core-backend/src/main/java/io/dataease/chart/server/ChartDataServer.java
@@ -309,7 +309,6 @@ public class ChartDataServer implements ChartDataApi {
             }
         } else {
             exportCenterManage.addTask(request.getViewId(), "chart", request);
-            return;
         }
     }
