+++ core/core-frontend/src/views/visualized/data/dataset/ExportExcel.vue
@@ -13,7 +13,8 @@ import {
   exportDelete,
   exportDeleteAll,
   exportDeletePost,
-  exportTasksRecords
+  exportTasksRecords,
+  generateDownloadUri
 } from '@/api/dataset'
 import { useI18n } from '@/hooks/web/useI18n'
 import { useEmitt } from '@/hooks/web/useEmitt'
@@ -220,12 +221,16 @@ const callbackExportSuc = () => {
 const downLoadAll = () => {
   if (multipleSelection.value.length === 0) {
     tableData.value.forEach(item => {
-      window.open(PATH_URL + '/exportCenter/download/' + item.id)
+      generateDownloadUri(item.id).then(res => {
+        window.open(PATH_URL + '/exportCenter/download/' + item.id)
+      })
     })
     return
   }
   multipleSelection.value.map(ele => {
-    window.open(PATH_URL + '/exportCenter/download/' + ele.id)
+    generateDownloadUri(ele.id).then(res => {
+      window.open(PATH_URL + '/exportCenter/download/' + ele.id)
+    })
   })
 }
 const showMsg = item => {
@@ -242,7 +247,9 @@ const timestampFormatDate = value => {
 import { PATH_URL } from '@/config/axios/service'
 import GridTable from '../../../../components/grid-table/src/GridTable.vue'
 const downloadClick = item => {
-  window.open(PATH_URL + '/exportCenter/download/' + item.id, openType)
+  generateDownloadUri(item.id).then(res => {
+    window.open(PATH_URL + '/exportCenter/download/' + item.id, openType)
+  })
 }
 
 const retry = item => {
